Update quiver hashcode dependency
Development on the quiver_* subpackages has been discontinued in favour
of the main quiver package at https://github.com/~/quiver-dart. The
code itself is identical, but is more actively maintained.

Since this package depends on Flutter, I've migrated instead to the hash
functions in dart:ui. This allows you to drop the dependency on quiver
entirely.